Split configs for submodules.
