+++ core/src/main/scala/org/apache/spark/internal/config/package.scala
@@ -656,85 +656,4 @@ package object config {
       .stringConf
       .toSequence
       .createWithDefault(Nil)
-
-  private[spark] object History {
-
-    val DEFAULT_LOG_DIR = "file:/tmp/spark-events"
-
-    val EVENT_LOG_DIR = ConfigBuilder("spark.history.fs.logDirectory")
-      .stringConf
-      .createWithDefault(DEFAULT_LOG_DIR)
-
-    val MAX_LOG_AGE_S = ConfigBuilder("spark.history.fs.cleaner.maxAge")
-      .timeConf(TimeUnit.SECONDS)
-      .createWithDefaultString("7d")
-
-    val LOCAL_STORE_DIR = ConfigBuilder("spark.history.store.path")
-      .doc("Local directory where to cache application history information. By default this is " +
-        "not set, meaning all history information will be kept in memory.")
-      .stringConf
-      .createOptional
-
-    val MAX_LOCAL_DISK_USAGE = ConfigBuilder("spark.history.store.maxDiskUsage")
-      .bytesConf(ByteUnit.BYTE)
-      .createWithDefaultString("10g")
-
-    val HISTORY_SERVER_UI_PORT = ConfigBuilder("spark.history.ui.port")
-      .doc("Web UI port to bind Spark History Server")
-      .intConf
-      .createWithDefault(18080)
-
-    val FAST_IN_PROGRESS_PARSING =
-      ConfigBuilder("spark.history.fs.inProgressOptimization.enabled")
-        .doc("Enable optimized handling of in-progress logs. This option may leave finished " +
-          "applications that fail to rename their event logs listed as in-progress.")
-        .booleanConf
-        .createWithDefault(true)
-
-    val END_EVENT_REPARSE_CHUNK_SIZE =
-      ConfigBuilder("spark.history.fs.endEventReparseChunkSize")
-        .doc("How many bytes to parse at the end of log files looking for the end event. " +
-          "This is used to speed up generation of application listings by skipping unnecessary " +
-          "parts of event log files. It can be disabled by setting this config to 0.")
-        .bytesConf(ByteUnit.BYTE)
-        .createWithDefaultString("1m")
-  }
-
-  private[spark] object Status {
-
-    val ASYNC_TRACKING_ENABLED = ConfigBuilder("spark.appStateStore.asyncTracking.enable")
-      .booleanConf
-      .createWithDefault(true)
-
-    val LIVE_ENTITY_UPDATE_PERIOD = ConfigBuilder("spark.ui.liveUpdate.period")
-      .timeConf(TimeUnit.NANOSECONDS)
-      .createWithDefaultString("100ms")
-
-    val MAX_RETAINED_JOBS = ConfigBuilder("spark.ui.retainedJobs")
-      .intConf
-      .createWithDefault(1000)
-
-    val MAX_RETAINED_STAGES = ConfigBuilder("spark.ui.retainedStages")
-      .intConf
-      .createWithDefault(1000)
-
-    val MAX_RETAINED_TASKS_PER_STAGE = ConfigBuilder("spark.ui.retainedTasks")
-      .intConf
-      .createWithDefault(100000)
-
-    val MAX_RETAINED_DEAD_EXECUTORS = ConfigBuilder("spark.ui.retainedDeadExecutors")
-      .intConf
-      .createWithDefault(100)
-
-    val MAX_RETAINED_ROOT_NODES = ConfigBuilder("spark.ui.dagGraph.retainedRootRDDs")
-      .intConf
-      .createWithDefault(Int.MaxValue)
-
-    val APP_STATUS_METRICS_ENABLED =
-      ConfigBuilder("spark.app.status.metrics.enabled")
-        .doc("Whether Dropwizard/Codahale metrics " +
-          "will be reported for the status of the running spark app.")
-        .booleanConf
-        .createWithDefault(false)
-  }
 }
